fix: Use config file REL path to tmp
This fixes the issue where /tmp/template-xxxx/template/waku[.ext] is
passed into the ignore rules, causing it to ignore everything and only a
LICENSE file is written. This patch just converts the path to the
relative path to the cloned dir in tmp. In other words the above turns
into template/waku[.ext].

Signed-off-by: AlexNg <contact@ngjx.org>